requirements.txt
PyQt5==5.15.10
pyqtgraph==0.13.3
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.8.2
